requirements.txt
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
numpy==1.26.4
opencv-python==4.9.0.80
facenet-pytorch==2.6.0
scikit-learn
pillow
matplotlib